cephadm: move more funcs into net_utils.py
